Add v2 data: calculator tool, grounding discipline, and the no-sheet slices
Five new slices, each traceable to a bucket in the error analysis rather than to
a hunch:

  tool_tax_bill / tool_matching   chained arithmetic (4.7% on tax bills) is now
                                  delegated to a calc tool. Rows carry a real
                                  call/result exchange, so the model learns to
                                  delegate rather than attempt the multiplication.
  grounding_discipline            the largest bucket: 68% of answers invented a
                                  figure, and only ~7% of those were derivable
                                  from anything provided. These reward naming few
                                  figures and reasoning qualitatively.
  no_sheet_enumeration            the degenerate loop -- bounded lists answered
                                  from the rules with no roster in sight.
  missing_data_request            its other half: asked about a team with nothing
                                  pasted, ask for the sheet instead of inventing.

capengine/calc.py evaluates arithmetic by walking a parsed AST; eval() is never
called, so an expression from a model -- or later from a stranger using the
playground -- cannot reach the filesystem or the interpreter. 11 tests cover both
the math and the refusals.

Also makes salary matching round rather than truncate. An odd salary times 1.25
lands on a half dollar, where the engine sat one cent below the tool; that
disagreement has no basis in the rules and would have taught inconsistent
arithmetic. No published figure changes.

3,500 v2 rows generated with zero verifier rejections; combined set 9,740 rows
(8,960 train / 780 eval, now including probes for the failure modes v1's eval
could not see). 218 tests pass.

+++ capengine/calc.py
@@ -110,7 +110,7 @@ def calc(expression: str) -> CalcResult:
 
     # Money lands on whole dollars; keep floats only when a genuine fraction survives.
     if isinstance(value, float) and abs(value - round(value)) < 1e-6:
-        value = int(round(value))
+        value = round(value)
     return CalcResult(expression=cleaned, value=value)
